# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type OperationalResponseReadiness = { [key: string]: unknown };

export type OperationalResponseRowsItem = { [key: string]: unknown };

export interface OperationalResponse {
  readiness: OperationalResponseReadiness;
  rows: OperationalResponseRowsItem[];
}

export interface DashboardMetric {
  label: string;
  value: string;
  delta: string;
}

export interface DashboardOpportunity {
  title: string;
  score: string;
  evidence: string;
  risk: string;
  state: string;
}

export type DashboardActivityTone = typeof DashboardActivityTone[keyof typeof DashboardActivityTone];


export const DashboardActivityTone = {
  verified: 'verified',
  ready: 'ready',
  approval: 'approval',
  normal: 'normal',
} as const;

export interface DashboardActivity {
  title: string;
  detail: string;
  result: string;
  tone: DashboardActivityTone;
}

export type DashboardSnapshotState = typeof DashboardSnapshotState[keyof typeof DashboardSnapshotState];


export const DashboardSnapshotState = {
  live: 'live',
  unavailable: 'unavailable',
} as const;

export type DashboardSnapshotEngine = {
  pagesAnalyzed: number;
  opportunities: number;
  actionsPrepared: number;
  executed: number;
  verified: number;
  regressions: number;
};

export type DashboardSnapshotVerification = {
  verified: number;
  total: number;
  pending: number;
  rolledBack: number;
  regressions: number;
};

export type DashboardSnapshotAiVisibility = {
  citationRate: string;
  brandMentionRate: string;
  citationShare: string;
};

export type DashboardSnapshotLearning = {
  signalCount: number;
  averageConfidence: string;
};

export type DashboardSnapshotImpact = {
  verifiedOptimizations: number;
  completedExperiments: number;
  rollbacks: number;
  regressionsDetected: number;
};

export interface DashboardSnapshot {
  state: DashboardSnapshotState;
  /** @nullable */
  reason: string | null;
  siteName: string;
  domain: string;
  generatedAt: string;
  dataFreshness: string;
  stale: boolean;
  approvalsPending: number;
  metrics: DashboardMetric[];
  engine: DashboardSnapshotEngine;
  opportunities: DashboardOpportunity[];
  activity: DashboardActivity[];
  verification: DashboardSnapshotVerification;
  aiVisibility: DashboardSnapshotAiVisibility;
  learning: DashboardSnapshotLearning;
  impact: DashboardSnapshotImpact;
}

export type GetDashboardParams = {
days?: GetDashboardDays;
country?: string;
device?: GetDashboardDevice;
};

export type GetDashboardDays = typeof GetDashboardDays[keyof typeof GetDashboardDays];


export const GetDashboardDays = {
  NUMBER_7: 7,
  NUMBER_28: 28,
  NUMBER_90: 90,
} as const;

export type GetDashboardDevice = typeof GetDashboardDevice[keyof typeof GetDashboardDevice];


export const GetDashboardDevice = {
  all: 'all',
  desktop: 'desktop',
  mobile: 'mobile',
  tablet: 'tablet',
} as const;

export type GetPerformanceParams = {
days?: GetPerformanceDays;
country?: string;
device?: GetPerformanceDevice;
};

export type GetPerformanceDays = typeof GetPerformanceDays[keyof typeof GetPerformanceDays];


export const GetPerformanceDays = {
  NUMBER_7: 7,
  NUMBER_28: 28,
  NUMBER_90: 90,
} as const;

export type GetPerformanceDevice = typeof GetPerformanceDevice[keyof typeof GetPerformanceDevice];


export const GetPerformanceDevice = {
  all: 'all',
  desktop: 'desktop',
  mobile: 'mobile',
  tablet: 'tablet',
} as const;

export type AskSeoEngineBody = {
  question: string;
};

export type AskSeoEngine200 = {
  answer?: string;
};