# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /performance:
    get:
      operationId: getPerformance
      parameters:
        - { name: days, in: query, schema: { type: integer, enum: [7, 28, 90] } }
        - { name: country, in: query, schema: { type: string } }
        - { name: device, in: query, schema: { type: string, enum: [all, desktop, mobile, tablet] } }
      responses:
        "200": { description: Performance data, content: { application/json: { schema: { $ref: "#/components/schemas/OperationalResponse" } } } }
...
  /command:
    post:
      operationId: askSeoEngine
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [question], properties: { question: { type: string } } } } }
      responses:
        "200": { description: Read-only answer, content: { application/json: { schema: { type: object, properties: { answer: { type: string } } } } } }
components:
  schemas:
...
    OperationalResponse:
      type: object
      properties:
        readiness: { type: object }
        rows: { type: array, items: { type: object, additionalProperties: true } }
      required: [readiness, rows]
...