# lib/api-zod/src/generated/types/operationalResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OperationalResponseReadiness } from './operationalResponseReadiness';
import type { OperationalResponseRowsItem } from './operationalResponseRowsItem';

export interface OperationalResponse {
  readiness: OperationalResponseReadiness;
  rows: OperationalResponseRowsItem[];
}